twilio/requirements.txt
Flask>=0.12
twilio~=6.0.0
python-dotenv=0.19.2